gestion des conversions en hexa, affichage adresse pointeur, début conversion int

+++ src/ft_convert.c
@@ -6,17 +6,17 @@
 /*   By: audumont <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/09 16:22:42 by audumont          #+#    #+#             */
-/*   Updated: 2019/11/09 20:08:14 by audumont         ###   ########.fr       */
+/*   Updated: 2019/11/10 14:34:26 by audumont         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../lib_printf.h"
 
-void		ft_convert(int nb, int base, char *tank)
+void		ft_convert(long long int nb, int base, char *tank)
 {
-	unsigned int nbr;
-	int		sign;
-	int		i;
+	long long int  nbr;
+	long long int		sign;
+	long long int		i;
 
 	nbr = nb & ((~0) >> 1);
 	sign = nb > 0 ? 1 : -1;
